Move properties/tags code from IndexerVerticle to MetadataVerticle
