+++ georocket-server/src/main/java/io/georocket/constants/AddressConstants.java
@@ -11,8 +11,8 @@ public final class AddressConstants {
   public static final String INDEXER_ADD = "georocket.indexer.add";
   public static final String INDEXER_QUERY = "georocket.indexer.query";
   public static final String INDEXER_DELETE = "georocket.indexer.delete";
-  public static final String INDEXER_UPDATE = "georocket.indexer.update";
   public static final String METADATA_GET_PROPERTIES = "georocket.metadata.get.properties";
+  public static final String METADATA_UPDATE = "georocket.metadata.update";
   public static final String ACTIVITIES = "georocket.activities";
   
   private AddressConstants() {
